ユーザーの詳細画面のスタイルを追加
